Emit Debug Event Start Of Def
Previously, emitting debugging events was restricted to:
1. the various junctures within polyfurcation expressions
2. the end of a `def`/`defp`

I deliberately desisted from including the starting of a `def`/`defp` out
of fear of redundancy. While using this library in my own project, it
became apparent that this event would greatly help to understand the linear
code path that is being taken from `function` to `function`. It also
dawned upon me that redundancy of events can be addressed on the `Debug
Viewer`-level.

This change required me to modify all the tests which although was kinda
painful, it gave me the opportunity to have the tests refactored for
better readibility.