Disable AST Annotation
When the option in `capture_medium` for `use ExDebugger` is set to `:none`,
then disable `AST` annotation altogether.

Before this change, I noticed that the codebase that I am beta testing
this on was running noticeably slower even though output was fully
disabled. By moving the utility of this setting to conditionally annotate
the AST instead I can assure that no performance penalties will be
incurred when a module does `use ExDebugger` with output disabled.

This feature is highly important as it allows `use ExDebugger` to
remain present in `prod`; closing the cognitive dissonance that
software developers tend to face between `prod` and `dev`.